SwissTypeFanuc -> HanwhaSwissType
