+++ Distributive/TurnMill/HanwhaSwissType/Registers.cs
@@ -12,6 +12,9 @@ namespace SprutTechnology.SCPostprocessor
         ///<summary>Automatic block numbers counter</summary>
         public CountingNCWord BlockN = new CountingNCWord("N{######}", 1, 1, 1);
 
+        ///<summary>Cycle block numbers counter</summary>
+        public NumericNCWord CycleBlockN = new NumericNCWord("N{######}",0);
+
         ///<summary>G90, G91 - absolute or incremental motion mode selection</summary>
         public NumericNCWord GAbsInc = new NumericNCWord("G{##}", 90);
 
@@ -55,7 +58,7 @@ namespace SprutTechnology.SCPostprocessor
         public NumericNCWord GHome = new NumericNCWord("G{##}", 28);
 
         ///<summary>G80, G81, G82... - canned cycle code</summary>
-        public NumericNCWord GCycle = new NumericNCWord("G{##}", 80);
+        public NumericNCWord GCycle = new NumericNCWord("G{##.#}", 80);
 
         ///<summary>G40, G41, G42 - radius compensation mode: off, left, right</summary>
         public NumericNCWord GRCompens = new NumericNCWord("G{##}", 40);
@@ -84,6 +87,8 @@ namespace SprutTechnology.SCPostprocessor
         ///<summary>H - incremental C axis movement</summary>
         public NumericNCWord H = new NumericNCWord("H{-#####!###}", 0);
 
+        public NumericNCWord P = new NumericNCWord("P{-#####!###}", 0);
+
         ///<summary>Circle radius</summary>
         public NumericNCWord R = new NumericNCWord("R{-#####!###}", 0);
 
@@ -100,7 +105,7 @@ namespace SprutTechnology.SCPostprocessor
         public NumericNCWord K = new NumericNCWord("K{-#####!###}", 0);
 
         ///<summary>Feedrate value of the movement</summary>
-        public NumericNCWord F = new NumericNCWord("F{#####}", 0);
+        public NumericNCWord F = new NumericNCWord("F{#####.##}", 0);
 
         ///<summary>Tool number (first two numbers)</summary>
         public NumericNCWord T = new NumericNCWord("T{00}", 0);
@@ -136,6 +141,7 @@ namespace SprutTechnology.SCPostprocessor
         {
             Block = new NCBlock(this, 
                 BlockN,
+                CycleBlockN,
                 GAbsInc,
                 GMeasure,
                 GWCS,
@@ -158,18 +164,19 @@ namespace SprutTechnology.SCPostprocessor
                 W,
                 H,
                 R,
+                P,
                 RSafeLevel,
                 I,
                 J,
                 K,
-                XDelay,
-                F,
+                XDelay,                
                 PDrillPause,
                 QStep,
                 QThreadAngle,
                 S,
                 T,
                 TCor,
+                F,
                 M,
                 MSpindle,
                 MCoolant,
