+++ Distributive/TurnMill/HanwhaSwissType/SyncPoints.cs
@@ -18,6 +18,7 @@ namespace SprutTechnology.SCPostprocessor
         Dictionary<int, int> fList;
         bool fSynchronized;
         Postprocessor fPost;
+        int step = 1;
 
         public TSyncPoints(Postprocessor post)
         {
@@ -31,27 +32,44 @@ namespace SprutTechnology.SCPostprocessor
                 if (scSyncPoint>=0) {
                     int pnt;
                     if (!fList.TryGetValue(scSyncPoint, out pnt)) {
-                        pnt = fLastPoint+1;
+                        pnt = fLastPoint + step;
                         fList.Add(scSyncPoint, pnt);
                     }
                     fPost.nc.WriteLine("M"+pnt);
                     fLastPoint = pnt;
                 } else {
-                    fLastPoint++;
+                    fLastPoint = fLastPoint + step;
                     fPost.nc1.WriteLine("M"+fLastPoint);
                     fPost.nc2.WriteLine("M"+fLastPoint);
                 }
                 if (fLastPoint>599) {
-                    Log.Error("Превышено максимальное количество точек синхронизации.");
+                    Log.Error("Exceeded the maximum number of wait marks.");
                 }
             }
             fSynchronized = true;
+        }
+        
+        public void OutPrev(int scSyncPoint)
+        {
+            int pnt;
+            if (fList.TryGetValue(scSyncPoint, out pnt)) {
+                pnt-= step;
+            }else{
+                fLastPoint += step;
+                pnt=fLastPoint;
+            }
+            fPost.nc.WriteLine("M"+pnt);
         }
 
         public void CheckForOutput(string s) 
         {
             if (!fSynchronized || String.IsNullOrEmpty(s) || s.StartsWith("("))   
                 return;
+            fSynchronized = false;
+        }
+
+        public void ResetSynchronized()
+        {
             fSynchronized = false;
         }
     }
