--- Distributive/TurnMill/SwissTypeFanuc/.vscode/launch.json
@@ -1,26 +1,0 @@
-{
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "Launch post",
-            "type": "coreclr",
-            "request": "launch",
-            "preLaunchTask": "buildPostprocessor",
-            "program": "${env:SprutCAMDir}/DotNet/InpCore.dll",
-            "args": ["-cldviewermode", "-sharedfile:STInpCoreOfDebugger", "-postfile:\"${workspaceFolder}\\bin\\Debug\\net5.0-windows\\${workspaceFolderBasename}.dll\""],
-            "cwd": "${workspaceFolder}",
-            "console": "integratedTerminal",
-            "stopAtEntry": false,
-            "justMyCode": true
-        },
-        {
-            "name": ".NET Core Attach",
-            "type": "coreclr",
-            "request": "attach",
-            "processId": "${command:pickProcess}"
-        }
-    ]
-}
