--- Distributive/TurnMill/SwissTypeFanuc/SwissTypeFanuc.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0-windows</TargetFramework>
-  </PropertyGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Settings.xml"/>
-    <Reference Include="STDefLib">
-      <HintPath>$(SprutCAMDir)\DotNet\STDefLib.dll</HintPath>      
-    </Reference>
-    <Reference Include="VecMatrLib">
-      <HintPath>$(SprutCAMDir)\DotNet\VecMatrLib.dll</HintPath>      
-    </Reference>
-    <Reference Include="SCPostprocessor">
-      <HintPath>$(SprutCAMDir)\DotNet\SCPostprocessor.dll</HintPath>      
-    </Reference>
-  </ItemGroup>
-
-</Project>
